Move apropriate functions to Job class
